Update CHANGELOG.md
Hello Taiko Team!

I'm new to the world of coding and GitHub, and I'm incredibly eager to learn everything about the wonderful universe of Taiko! 😊 While following Taiko's development, I noticed something in your changelog that caught my attention, and I thought I'd share it with you.

In your changelog, under the heading "implement release-please workflow," I noticed that unlike other entries, this particular update doesn't start with a colon. In other entries, updates are typically presented in a "heading: description" format. I was wondering if this was an intentional stylistic choice or perhaps an oversight?

If it's a slight oversight, adjusting it to match the format of other entries might make the changelog more consistent and even easier for beginners like me to follow. Of course, this is just a suggestion. 😊

Following Taiko's journey and being a part of this amazing project is incredibly exciting for me. I'm looking forward to growing and learning with you!

Thank you, and have a fantastic day!